refactor(elysia): nest forum/traces/schedule routes (one-route-per-file)
Restructure flat forum.ts / traces.ts / schedule.ts into per-module
directories with one endpoint per file, mirroring the test layout.

- forum/{index,model,threads-list,thread-create,thread-get,thread-status}.ts
- traces/{index,model,list,get,chain,link,unlink,linked-chain}.ts
- schedule/{index,model,md,list,create,update}.ts

Built-in TypeBox (`elysia.t`) only — no zod. Each file ≤40 lines.
