+++ src/routes-elysia/forum/index.ts
@@ -1,0 +1,11 @@
+import { Elysia } from 'elysia';
+import { threadsListRoute } from './threads-list.ts';
+import { threadCreateRoute } from './thread-create.ts';
+import { threadGetRoute } from './thread-get.ts';
+import { threadStatusRoute } from './thread-status.ts';
+
+export const forumApi = new Elysia()
+  .use(threadsListRoute)
+  .use(threadCreateRoute)
+  .use(threadGetRoute)
+  .use(threadStatusRoute);
